+++ client/src/renderers/handlebars.js
@@ -5,6 +5,7 @@ var handlebars = require('handlebars');
 module.exports = function(){
 
   this.render = function(template, model, callback){
+
     var linked = handlebars.template(template),
         html = linked(model);