+++ client/src/settings.js
@@ -6,6 +6,7 @@ module.exports = {
   connectionError: '{0} when connecting to {1}',
   emptyResponse: 'Empty response',
   genericError: 'Internal client error',
+  legacyComponent: 'The component can\'t be rendered because it was published with an older OC version',
   registriesEmpty: 'registries must contain at least one endpoint',
   registriesIsNotObject: 'registries must be an object',
   serverSideRenderingFail: 'Server-side rendering failed: {0}'