+++ src/cli/domain/get-components-deps.js
@@ -5,12 +5,25 @@ var path = require('path');
 var _ =  require('underscore');
 
 module.exports = function(components){
-  var deps = [];
+
+  var deps = { modules: [], withVersions: [] };
+
   _.forEach(components, function(c){
+
     var pkg = fs.readJsonSync(path.join(c, 'package.json'));
+
     _.forEach(_.keys(pkg.dependencies), function(d){
-      if(!_.contains(deps, d)){
-        deps.push(d);
+
+      var version = pkg.dependencies[d],
+          hasVersion = !_.isEmpty(version),
+          depToInstall = hasVersion ? (d + '@' + version) : d;
+
+      if(!_.contains(deps.withVersions, depToInstall)){
+        deps.withVersions.push(depToInstall);
+      }
+
+      if(!_.contains(deps.modules, d)){
+        deps.modules.push(d);
       }
     });
   });