+++ src/cli/domain/npm-installer.js
@@ -3,9 +3,9 @@
 var npm = require('npm');
 var path = require('path');
 
-module.exports = function(dependencies, baseDir, cb){
+module.exports = function(dependencies, cb){
   npm.load({}, function(npmEr){
     if(!!npmEr){ return cb(npmEr); }
-    npm.commands.install(path.resolve(baseDir), dependencies, cb);
+    npm.commands.install(path.resolve('.'), dependencies, cb);
   });
 };