+++ src/components/base-component-handlebars/_package/template.js
@@ -1,1 +1,1 @@
-var oc=oc||{};oc.components=oc.components||{},oc.components.a34cffd0625f1eb8a08569f068cd4a9768104a91={compiler:[6,">= 2.0.0-beta.1"],main:function(o,c,n,a){return"  "},useData:!0};
+var oc=oc||{};oc.components=oc.components||{},oc.components["737b5c8ac8536a42972facd7f2bbfc945326df42"]={compiler:[7,">= 4.0.0"],main:function(c,o,n,a,e){return"  "},useData:!0};