+++ src/registry/domain/get-package-json-from-temp-dir.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var fs = require('fs-extra');
+var path = require('path');
+
+module.exports = function(tempDirPath, callback){
+  return fs.readJson(path.join(tempDirPath, 'package.json'), callback);
+};