+++ src/registry/domain/options-sanitiser.js
@@ -16,6 +16,12 @@ module.exports = function(input){
     options.beforePublish = auth.middleware(options.publishAuth);
   }
 
+  if(!options.publishValidation){
+    options.publishValidation = function(){
+      return { isValid: true };
+    };
+  }
+
   if(!options.prefix){
     options.prefix = '/';
   }
@@ -26,6 +32,10 @@ module.exports = function(input){
 
   if(!!options.dependencies){
     options.dependencies = dependenciesResolver(options);
+  }
+
+  if(!_.isBoolean(options.hotReloading)){
+    options.hotReloading = !!options.local;
   }
 
   return options;