+++ test/front-end/render.js
@@ -1,6 +1,7 @@
 'use strict';
 
-var handlebarsCompiledView = 'oc.components=oc.components||{},oc.components["46ee85c314b371cac60471cef5b2e2e6c443dccf"]={compiler:[6,">= 2.0.0-beta.1"],main:function(){return"Hello world!"},useData:!0};';
+var handlebars3CompiledView = 'oc.components=oc.components||{},oc.components["46ee85c314b371cac60471cef5b2e2e6c443dccf"]={compiler:[6,">= 2.0.0-beta.1"],main:function(){return"Hello world!"},useData:!0};';
+var handlebarsCompiledView = 'oc.components=oc.components||{},oc.components["46ee85c314b371cac60471cef5b2e2e6c443dccf"]={compiler:[7,">= 4.0.0"],main:function(){return"Hello world!"},useData:!0};';
 var jadeCompiledView = 'oc.components=oc.components||{},oc.components["09227309bca0b1ec1866c547ebb76c74921e85d2"]=function(n){var e,o=[],c=n||{};return function(n){o.push("<span>hello "+jade.escape(null==(e=n)?"":e)+"</span>")}.call(this,"name"in c?c.name:"undefined"!=typeof name?name:void 0),o.join("")};';
 
 describe('oc-client : render', function(){
@@ -29,10 +30,11 @@ describe('oc-client : render', function(){
   });
   
   describe('when rendering handlebars component', function(){
-    
+   
     describe('when handlebars runtime not loaded', function(){
 
       var originalHandlebars, originalHeadLoad, callback, headSpy;
+
       beforeEach(function(){
         originalHandlebars = Handlebars;
         originalHeadLoad = head.load;
@@ -62,7 +64,7 @@ describe('oc-client : render', function(){
 
       it('should require and wait for it', function(){
         expect(headSpy.called).toBe(true);
-        expect(headSpy.args[0][0]).toEqual('https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/3.0.1/handlebars.runtime.js');
+        expect(headSpy.args[0][0]).toEqual('https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.5/handlebars.runtime.min.js');
       });
 
       it('should render the component', function(){
@@ -101,7 +103,31 @@ describe('oc-client : render', function(){
       });
     });
   });
-  
+
+  describe('when handlebars runtime loaded and rendering a handlebars3 component', function(){
+
+    var callback, originalConsolelog = console.log;
+    beforeEach(function(){
+      console.log = function(){};
+      callback = sinon.spy();
+      eval(handlebars3CompiledView);
+      oc.render({
+        src: 'https://my-cdn.com/components/a-component/1.2.123/template.js', 
+        type: 'handlebars', 
+        key: '46ee85c314b371cac60471cef5b2e2e6c443dccf'
+      }, {}, callback);
+    });
+
+    afterEach(function(){
+      console.log = originalConsolelog;
+    });
+
+    it('should return the error', function(){ console.log(callback.args);
+      expect(callback.called).toBe(true);
+      expect(callback.args[0][0]).toContain('Template was precompiled with an older version of Handlebars than the current runtime');
+    });
+  });
+
   describe('when rendering jade component', function(){
     
     describe('when jade runtime not loaded', function(){